add to Makefile
the problem is, when using the template with the template button, it replicates the whole repo, then it triggers github actions. Github actions don't have permissions to edit/add github actions. This is a fundamental feature to stop infinite loops and bad acting. This means that the original ci.yml and codeql.yml have to be identical when to the /project_template version as it cannot then edit those files. This means that if the user chooses pipenv, as the repo uses poetry, it will fail.